improving layout, adding translations

+++ src/headless/shared/chatbox.js
@@ -3,7 +3,7 @@ import api from './api/index.js';
 import { isUniView } from '../utils/session.js';
 import _converse from './_converse.js';
 import converse from './api/public.js';
-import log from "@converse/log";
+import log from '@converse/log';
 import ModelWithMessages from './model-with-messages.js';
 
 const { u } = converse.env;
@@ -24,7 +24,7 @@ export default class ChatBoxBase extends ModelWithMessages(Model) {
             // but we're in embedded mode.
             return;
         }
-        this.set({'box_id': `box-${jid}`});
+        this.set({ 'box_id': `box-${jid}` });
     }
 
     validate(attrs) {
@@ -55,34 +55,50 @@ export default class ChatBoxBase extends ModelWithMessages(Model) {
                 // We only have one chat visible at any one time.
                 // So before opening a chat, we make sure all other chats are hidden.
                 other_chats.forEach((c) => u.safeSave(c, { hidden: true }));
-                u.safeSave(this, { hidden: false });
+                u.safeSave(this, { hidden: false, closed: false });
                 this.trigger('show');
             }
             return this;
         }
         // Overlayed view mode
-        u.safeSave(this, { 'hidden': false });
+        u.safeSave(this, { hidden: false, closed: false });
         this.trigger('show');
         return this;
+    }
+
+    async shouldDestroyOnClose() {
+        /**
+         * *Hook* which allows plugins to determine whether a chat should be destroyed when it's closed.
+         * For example, used by the converse-dragresize plugin to prevent resized chats
+         * from being destroyed, thereby losing the resize dimensions.
+         * @event _converse#shouldDestroyOnClose
+         * @param {ChatBoxBase} chatbox
+         * @param {boolean} should_destroy
+         */
+        return await api.hook('shouldDestroyOnClose', this, true);
     }
 
     /**
      * @param {Object} [_ev]
      */
     async close(_ev) {
-        try {
-            await new Promise((success, reject) => {
-                return this.destroy({
-                    success,
-                    error: (_m, e) => reject(e),
+        if (await this.shouldDestroyOnClose()) {
+            try {
+                await new Promise((success, reject) => {
+                    return this.destroy({
+                        success,
+                        error: (_m, e) => reject(e),
+                    });
                 });
-            });
-        } catch (e) {
-            log.debug(e);
-        } finally {
-            if (api.settings.get('clear_messages_on_reconnection')) {
-                await this.clearMessages();
+            } catch (e) {
+                log.debug(e);
             }
+        } else {
+            u.safeSave(this, { closed: true });
+        }
+
+        if (api.settings.get('clear_messages_on_reconnection')) {
+            await this.clearMessages();
         }
         /**
          * Triggered once a chatbox has been closed.
